backend/fastapi/requirements.txt
fastapi
uvicorn
pyjwt
